% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_pooling.R
\name{linsilver.mle.test}
\alias{linsilver.mle.test}
\title{Lin and Stiver's MLE based test statistic under heteroscedasticity}
\usage{
linsilver.mle.test(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{y}{a (non-empty) numeric vector of data values}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}
}
\value{
a list with class "htest" containing the below:\tabular{ll}{
   \code{statistic} \tab the value of the MLE based statistic \cr
   \tab \cr
   \code{p.value} \tab the p.value for the statistic \cr
   \tab \cr
   \code{alternative} \tab the alternative hypothesis for the test \cr
   \tab \cr
   \code{method} \tab indicates what type of pooling approach was performed \cr
   \tab \cr
   \code{data.name} \tab the names of the data \cr
}
}
\description{
Calculates MLE based test statistic and corresponding p-value based on a modified maximum likelihood estimator and simple mean difference for testing
the equality of two correlated means with incomplete data under the bivariate Gaussian assumption.
}
\examples{
# Using colondat sample dataset
data("colondat")  
linsilver.mle.test(colondat$`araB 1bn`,colondat$`araB 1bT`,alternative="two.sided")
}
\references{
Kuan, P., & Huang, B. (2013). A simple and robust method for partially matched samples using the p-values pooling approach. Statistics In Medicine, 32(19), 3247-3259. doi: 10.1002/sim.5758
}
