% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_pooling.R
\name{looneyjones.z.test}
\alias{looneyjones.z.test}
\title{Looney and Jones corrected Z-test}
\usage{
looneyjones.z.test(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{y}{a (non-empty) numeric vector of data values}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}
}
\value{
a list with class "htest" containing the below:\tabular{ll}{
   \code{statistic} \tab the value of the corrected Z-test \cr
   \tab \cr
   \code{p.value} \tab the p.value for the statistic \cr
   \tab \cr
   \code{alternative} \tab the alternative hypothesis for the test \cr
   \tab \cr
   \code{method} \tab indicates what type of pooling approach was performed \cr
   \tab \cr
   \code{data.name} \tab the names of the data \cr
}
}
\description{
Calculates corrected Z-test and corresponding p-value based on a modified variance estimation of the standard Z-test by accounting for the correlation among the n1 matched pairs.
The corrected Z-test reduces to paired sample or two-sample Z-test when n2=n3=0 or n1=0, respectively.
}
\examples{
# Using colondat sample dataset
data("colondat") 
looneyjones.z.test(colondat$`araB 1bn`,colondat$`araB 1bT`,alternative="two.sided")
}
\references{
Kuan, P., & Huang, B. (2013). A simple and robust method for partially matched samples using the p-values pooling approach. Statistics In Medicine, 32(19), 3247-3259. doi: 10.1002/sim.5758
}
