% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_pooling.R
\name{p.values.pooling}
\alias{p.values.pooling}
\title{P-value Pooling Strategies for Partially Matched Samples}
\usage{
p.values.pooling(x, y, method = "liptak", alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{y}{a (non-empty) numeric vector of data values}

\item{method}{a character string specifying which pooling method to perform, must be one of "liptak" (default), "kim", "looneyjones", "linsilver" or "ekbohm"}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}
}
\value{
a list with class "htest" containing the below:\tabular{ll}{
   \code{statistic} \tab the value of chosen method's statistic \cr
   \tab \cr
   \code{p.value} \tab the p.value for the statistic \cr
   \tab \cr
   \code{method} \tab indicates what type of pooling approach was performed \cr
   \tab \cr
   \code{data.name} \tab the names of the data \cr
}
}
\description{
Calculates the test statistic and corresponding p-value for partially matched samples based on the user's preferred pooling method.
Partially matched samples are viewed as data generated from two experimental designs; n1 matched pairs, and independent groups with n2 and n3 observations
per group.
}
\examples{
# Using colondat sample dataset
data("colondat") 
p.values.pooling(colondat$`araB 1bn`,colondat$`araB 1bT`,method="liptak",alternative="two.sided")
}
\references{
Kuan, P., & Huang, B. (2013). A simple and robust method for partially matched samples using the p-values pooling approach. Statistics In Medicine, 32(19), 3247-3259. doi: 10.1002/sim.5758
}
